backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
langchain==0.3.23
langgraph==0.3.31
requests==2.32.3
pillow==11.1.0
python-multipart==0.0.20
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.11.3
pydantic-settings==2.8.1
torch==2.7.0
transformers==4.51.3
safetensors==0.5.3
